Major JSON overhaul
Main difference is that the JSON type no longer processes every get as a string but rather as either a string or an int, and the original get is replaced into a repeatedGet and a get function. Also added some extra functionality such as isEmpty(), isNull() and length(). Should be faster now, as the inner nlohmann::json json type is now a pointer. Also removed one test as it didn't use the mock for some reason. Comments are not correct at the moment.